Added ignore path feature / New stalled issues cases added

+++ src/MEGASync/control/Utilities.h
@@ -313,7 +313,7 @@ public:
     static void getPROurlWithParameters(QString &url);
     static QString joinLogZipFiles(mega::MegaApi *megaApi, const QDateTime *timestampSince = nullptr, QString appendHashReference = QString());
 
-    static QString getElidedPath(const QString& path, int firstPart, int secondPart, int maxDepth);
+    static QString getElidedPath(const QString& path, uint8_t firstPart, uint8_t secondPart, uint8_t maxDepth);
 
     static void adjustToScreenFunc(QPoint position, QWidget *what);
     static QString minProPlanNeeded(std::shared_ptr<mega::MegaPricing> pricing, long long usedStorage);
